third_party/WebKit/LayoutTests/compositing/squashing/dont-squash-into-blend-mode.html

<!doctype HTML>
Any errors will show below this line.
<div id="target" style="position: absolute; width: 200px; height: 200px; will-change: transform; background: lightblue; mix-blend-mode: overlay;"></div>
<div style="position: absolute; width: 200px; height: 200px; top: 100px; left: 100px; background: lightgray"></div>
<script src="../../resources/testharness.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
onload = function() {
    if (window.internals) {
        var layers = JSON.parse(internals.layerTreeAsText(document, 1));
        assert_true(layers.children[0].children[0].children[1].compositingReasons[1] == "Squashing a layer with blending is not supported.");
    }
};
</script>
